Fix @authorize to authorize

+++ app/controllers/favorites_controller.rb
@@ -2,7 +2,7 @@ class FavoritesController < ApplicationController
   def create
     @post = Post.find(params[:post_id])
     favorite = current_user.favorites.build(post: @post)
-    @authorize favorite
+    authorize favorite
     if favorite.save
       flash[:notice] = "Success favoriting the post."
       redirect_to [@post.topic, @post]
@@ -15,7 +15,7 @@ class FavoritesController < ApplicationController
   def destroy
     @post = Post.find(params[:post_id])
     favorite = current_user.favorites.find_by_post_id(params[:post_id])
-    @authorize favorite
+    authorize favorite
     if favorite.destroy
       flash[:notice] = "Post unfavorited."
       redirect_to [@post.topic, @post]
